changed impl

+++ SW07/Calculator/src/main/java/hslu/oop/sw06/Calculator.java
@@ -6,7 +6,7 @@ public interface Calculator {
      * addition method for calculating two integers together
      * @param a
      * @param b
-     * @return sum of a and b
+     * @return sum of a and b as a long
      */
-    int addition(int a, int b);
+    long addition(int a, int b);
 }
