+++ SW07/Calculator/src/main/java/hslu/oop/sw06/CalculatorImpl.java
@@ -3,7 +3,7 @@ package hslu.oop.sw06;
 public class CalculatorImpl implements Calculator{
 
     @Override
-    public int addition(int a, int b) {
-        return a + b;
+    public long addition(int a, int b) {
+        return (long) a + b;
     }
 }
